proper handling of cognito error
