+++ src/runtime/defs.ts
@@ -3,3 +3,31 @@ export const PathAttributeName: string = '__path__';
 export const PathAttributeNameQuery: string = '__path__?';
 export const ParentAttributeName: string = '__parent__';
 export const DeletedFlagAttributeName: string = '__is_deleted__';
+
+export type UnautInfo = {
+  opr: string;
+  entity: string;
+};
+
+function asUnauthMessage(obj: string | UnautInfo): string {
+  if (typeof obj == 'string') {
+    return obj;
+  } else {
+    return `User not authorised to perform '${obj.opr}' on ${obj.entity}`;
+  }
+}
+
+export class UnauthorisedError extends Error {
+  constructor(message?: string | UnautInfo, options?: ErrorOptions) {
+    super(
+      message ? asUnauthMessage(message) : 'User not authorised to perform this operation',
+      options
+    );
+  }
+}
+
+export class BadRequestError extends Error {
+  constructor(message?: string, options?: ErrorOptions) {
+    super(message ? asUnauthMessage(message) : 'BadRequest', options);
+  }
+}
